requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
attrs==23.2.0
bcrypt==4.2.0
boltons==21.0.0
bracex==2.5
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-option-group==0.5.6
codebleu==0.7.0
colorama==0.4.6
cryptography==43.0.0
datasets
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
face==22.0.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.5.0
glom==22.1.0
googleapis-common-protos==1.63.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
hydra-core==1.3.2
idna==3.7
importlib_metadata==7.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.16
natsort==8.4.0
numpy
omegaconf==2.3.0
openai
openpyxl
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson
packaging==24.1
pandas
paramiko==3.4.0
peewee==3.17.6
pillow
portalocker==2.10.1
protobuf
pyarrow==17.0.0
pyarrow-hotfix==0.6
pycparser==2.22
pydantic
pydantic_core==2.20.1
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv
pytz==2024.1
PyYAML
referencing==0.35.1
regex
requests
rich==13.7.1
rpds-py==0.20.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
ruff==0.5.6
sacrebleu==2.4.2
semgrep==1.84.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
together==1.2.5
tomli==2.0.1
tqdm
tree-sitter
tree-sitter-c==0.21.4
tree-sitter-c-sharp==0.21.3
tree-sitter-cpp==0.22.3
tree-sitter-go==0.21.0
tree-sitter-java==0.21.0
tree-sitter-javascript==0.21.4
tree-sitter-php==0.22.8
tree-sitter-python==0.21.0
tree-sitter-ruby==0.21.0
tree-sitter-rust==0.21.2
typer==0.12.3
typing_extensions
tzdata==2024.1
urllib3==2.2.2
vt-py==0.18.3
wcmatch==8.5.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2

torch
evaluate
termcolor
transformers
Jinja2
markdown2
sympy
nltk
anthropic